GC fix: Copy instead of ref for popping elements.
Yes,  a single missing ampersand(&)
caused all this problem. It caused a test program for 'for' loop to get
stuck after a few iterations due to maybe too many environments in the GC.

We removed that timer too from the GC, becauses for reasons unknown it
caused a lot of pagefaults.